The whole "Chat" project
Server side is in the file Server.java
Client side is in the file Client.java
.jar file is titled Chat.jar
mainDB.db is a database containing a list of clients which may be authorized

There are a total of 10 records, each one is of the form "# login<#> password<#> nickname<#>", where # from 0 through 9 and <> are not needed. The first record is "0 login password nickname", the second "1 login1 password1 nickname1", and so on.
Please enter one of these records when asked to authorize, and you are up and running to exchange messages with the other clients.